Move translate classes into translate_d_sw
