+++ tests/translate/__init__.py
@@ -4,7 +4,7 @@ from .parallel_translate import ParallelTranslate
 from .translate_a2b_ord4 import TranslateA2B_Ord4
 from .translate_circulation_cgrid import TranslateCirculation_Cgrid
 from .translate_copycorners import TranslateCopyCorners
-from .translate_d_sw import TranslateD_SW
+from .translate_d_sw import TranslateD_SW, TranslateUbKE, TranslateVbKE
 from .translate_del6vtflux import TranslateDel6VtFlux
 from .translate_delnflux import TranslateDelnFlux, TranslateDelnFlux_2
 from .translate_divergencedamping import TranslateDivergenceDamping
@@ -26,9 +26,7 @@ from .translate_haloupdate import (
 )
 from .translate_ke_c_sw import TranslateKE_C_SW
 from .translate_pgradc import TranslatePGradC
-from .translate_ubke import TranslateUbKE
 from .translate_updatedzc import TranslateUpdateDzC
-from .translate_vbke import TranslateVbKE
 from .translate_vorticityvolumemean import TranslateVorticityVolumeMean
 from .translate_wdivergence import TranslateWdivergence
 from .translate_xppm import TranslateXPPM, TranslateXPPM_2
