+++ tests/translate/translate_d_sw.py
@@ -1,6 +1,7 @@
 from .translate import TranslateFortranData2Py
 import fv3core.stencils.d_sw as d_sw
 import fv3core.utils.gt4py_utils as utils
+from fv3core.stencils.d_sw import ubke, vbke
 
 
 class TranslateD_SW(TranslateFortranData2Py):
@@ -62,3 +63,53 @@ class TranslateD_SW(TranslateFortranData2Py):
 
         d_sw.compute(**inputs)
         return self.slice_output(inputs)
+
+
+class TranslateUbKE(TranslateFortranData2Py):
+    def _call(self, *args, **kwargs):
+        ubke(
+            *args,
+            **kwargs,
+            cosa=self.grid.cosa,
+            rsina=self.grid.rsina,
+            origin=self.grid.compute_origin(),
+            domain=self.grid.domain_shape_compute_buffer_2d(),
+            splitters=self.grid.splitters,
+        )
+
+    def __init__(self, grid):
+        super().__init__(grid)
+        self.compute_func = self._call
+        self.in_vars["data_vars"] = {
+            "uc": {},
+            "vc": {},
+            "ut": {},
+            "ub": grid.compute_dict_buffer_2d(),
+        }
+        self.in_vars["parameters"] = ["dt5", "dt4"]
+        self.out_vars = {"ub": grid.compute_dict_buffer_2d()}
+
+
+class TranslateVbKE(TranslateFortranData2Py):
+    def _call(self, *args, **kwargs):
+        vbke(
+            *args,
+            **kwargs,
+            cosa=self.grid.cosa,
+            rsina=self.grid.rsina,
+            origin=self.grid.compute_origin(),
+            domain=self.grid.domain_shape_compute_buffer_2d(),
+            splitters=self.grid.splitters,
+        )
+
+    def __init__(self, grid):
+        super().__init__(grid)
+        self.compute_func = self._call
+        self.in_vars["data_vars"] = {
+            "vc": {},
+            "uc": {},
+            "vt": {},
+            "vb": grid.compute_dict_buffer_2d(),
+        }
+        self.in_vars["parameters"] = ["dt5", "dt4"]
+        self.out_vars = {"vb": grid.compute_dict_buffer_2d()}
